+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.6.1] - 2024-06-22
 
 ### Changed
 
@@ -83,6 +83,7 @@
 - `sync` subcommand support to sync with latest Tinted Theming schemes
 - `build` subcommand to trigger theme template build
 
+[0.6.1]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.3.0...v0.4.0